Звезды #v.0.9.7.rev.A.build.5

+++ core/src/ru/asupd/poop_ballon/Workers/Resizer.java
@@ -13,7 +13,7 @@ public class Resizer {
     int position=0;
     float current_dt=0;
     boolean start=false;
-    public Resizer(int x,int y) {
+    public Resizer(float x, float y) {
         size_x=x;
         size_y=y;
     }
